Add iNes Mapper 003 (CNROM) Support
Add support for iNES Mapper 003, which is a very simple mapper
only swiching out CHR chunks.

There is 16kB of fixed space mapped from $C000-FFFF and,
optionally 16kB more of fixed space from $8000-$BFFF (if not,
this area is mirrored)

The CHR ROM from $0000-$1FFF is banked in 8k chunks, up to four.
Some games look for bus contention, it is not implemented.